requirements.txt
Django==4.1.4
djangorestframework==3.14.0
drf-spectacular==0.25.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
drf-nested-routers
django-filter